+++ RELEASE_NOTES.md
@@ -14,6 +14,7 @@ Changes:
   * Added workstations
   * Added thermal cycler instrument type
   * Allow recording the workstation and thermal cycler used to prepare libraries
+  * Allow moving box to another location from the Edit Transfer page Receipt Wizard
   * Added another version of the Index Distance tool to the MISO online documentation site
   * Fixed an issue where Run Scanner could cause a container to repeatedly flip between multiple
     container models if the container is linked to multiple runs